[CLOUD-2777]: Adds schema caching and download
Implements caching and downloading of schemas to improve performance and reduce reliance on external resources.

This change introduces a new cache for schemas, separate from the mage runner, and includes a script to download schemas from a specified S3 bucket if the cache is not hit. This allows for faster access to schemas and reduces the need to fetch them from S3 every time.